Simplify splitting :provider into :provider-plugins
by letting :provider-plugins depend directly on :provider
and removing :provider-spi
