--- provider-spi/build.gradle.kts
@@ -1,14 +1,0 @@
-import build.*
-
-plugins {
-    id("public-kotlin-dsl-module")
-}
-
-base {
-    archivesBaseName = "gradle-kotlin-dsl-provider-spi"
-}
-
-dependencies {
-    compileOnly(gradleApi())
-    compile(futureKotlin("stdlib-jdk8"))
-}
