+++ provider/src/main/kotlin/org/gradle/kotlin/dsl/accessors/CodeGenerator.kt
@@ -18,15 +18,11 @@ package org.gradle.kotlin.dsl.accessors
 
 import org.gradle.api.internal.HasConvention
 import org.gradle.api.plugins.ExtensionAware
-import org.gradle.kotlin.dsl.provider.spi.ProjectSchema
-import org.gradle.kotlin.dsl.provider.spi.ProjectSchemaEntry
-import org.gradle.kotlin.dsl.provider.spi.TypeAccessibility
 
 import org.jetbrains.kotlin.lexer.KotlinLexer
 import org.jetbrains.kotlin.lexer.KtTokens
 
 
-internal
 fun ProjectSchema<TypeAccessibility>.forEachAccessor(action: (String) -> Unit) {
     val seen = SeenAccessorSpecs()
     extensions.mapNotNull(::typedAccessorSpec).forEach { spec ->
