+++ provider/src/main/kotlin/org/gradle/kotlin/dsl/provider/KotlinScriptTarget.kt
@@ -30,15 +30,14 @@ import org.gradle.internal.classpath.ClassPath
 import org.gradle.kotlin.dsl.KotlinBuildScript
 import org.gradle.kotlin.dsl.KotlinInitScript
 import org.gradle.kotlin.dsl.KotlinSettingsScript
+import org.gradle.kotlin.dsl.accessors.AccessorsClassPath
 import org.gradle.kotlin.dsl.accessors.accessorsClassPathFor
-import org.gradle.kotlin.dsl.provider.spi.AccessorsClassPath
-import org.gradle.kotlin.dsl.provider.spi.KotlinScriptBasePluginsApplicator
-import org.gradle.kotlin.dsl.provider.spi.serviceOf
 import org.gradle.kotlin.dsl.support.KotlinBuildscriptBlock
 import org.gradle.kotlin.dsl.support.KotlinInitscriptBlock
 import org.gradle.kotlin.dsl.support.KotlinPluginsBlock
 import org.gradle.kotlin.dsl.support.KotlinScriptHost
 import org.gradle.kotlin.dsl.support.KotlinSettingsBuildscriptBlock
+import org.gradle.kotlin.dsl.support.serviceOf
 
 import java.lang.IllegalArgumentException
