+++ provider/src/main/kotlin/org/gradle/kotlin/dsl/services/KotlinScriptServiceRegistry.kt
@@ -33,7 +33,6 @@ class KotlinScriptServiceRegistry : AbstractPluginServiceRegistry() {
     }
 
     override fun registerGradleUserHomeServices(registration: ServiceRegistration) {
-        registration.addProvider(org.gradle.kotlin.dsl.accessors.GradleUserHomeServices)
         registration.addProvider(org.gradle.kotlin.dsl.support.GradleUserHomeServices)
         registration.addProvider(org.gradle.kotlin.dsl.provider.GradleUserHomeServices)
     }
