+++ provider/src/test/kotlin/org/gradle/kotlin/dsl/accessors/GenerateProjectSchemaTest.kt
@@ -1,9 +1,5 @@
 package org.gradle.kotlin.dsl.accessors
 
-import org.gradle.kotlin.dsl.provider.spi.ProjectSchema
-import org.gradle.kotlin.dsl.provider.spi.ProjectSchemaEntry
-import org.gradle.kotlin.dsl.provider.spi.loadMultiProjectSchemaFrom
-
 import org.gradle.kotlin.dsl.fixtures.AbstractIntegrationTest
 
 import org.hamcrest.CoreMatchers.equalTo
