+++ settings.gradle.kts
@@ -4,7 +4,6 @@ rootProject.name = "gradle-kotlin-dsl"
 
 include(
     "provider",
-    "provider-spi",
     "provider-plugins",
     "tooling-models",
     "tooling-builders",
